MonsterDesc almost done
