+++ src/Engine/Objects/Monsters.h
@@ -6,6 +6,9 @@
 #include <vector>
 
 #include "Utility/Blob.h"
+#include "Utility/IndexedArray.h"
+
+#include "ActorEnums.h"
 
 /*  334 */
 enum MONSTER_TYPE {
@@ -209,9 +212,9 @@ struct MonsterDesc {
     uint16_t uMovementSpeed;
     int16_t uToHitRadius;
     uint32_t sTintColor;
-    uint16_t pSoundSampleIDs[4];
+    std::array<uint16_t, 4> pSoundSampleIDs; // TODO: actually this one should also be an IndexedArray
     std::string pMonsterName;
-    std::array<std::string, 10> pSpriteNames;
+    IndexedArray<std::string, ANIM_Count> pSpriteNames;
 };
 #pragma pack(pop)
